requirements.txt
beautifulsoup4==4.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
